Receipt multiple usage example (#26)
Introduces a credit-based purchasing scheme where a single receipt can be used multiple times until all credits associated with it have been consumed. In this case the receipt is used to generate images of 19th century US presidents.

Running the example from the buyer-side:

pnpm run images

This example first hits the /api/images/buy endpoint to get a PRT that it can use to purchase image-generation credits. The buyer then pays the bill and hits /api/images/generate multiple times with the Receipt it got back; each time the seller deducts the credits attached to that receipt by 1.